Correcao migrations

+++ questoes_vf/db/migrate/20130314194455_create_proposicaos.rb
@@ -3,7 +3,7 @@ class CreateProposicaos < ActiveRecord::Migration
     create_table :proposicaos do |t|
       t.references :questao
       t.boolean :resposta
-      t.double :valor
+      t.decimal :valor
 
       t.timestamps
     end
